Making cluster understand itself: matches up running instances with non
