+++ clusters/demohadoop.rb
@@ -8,7 +8,7 @@
 #        and dispatch those to the chef server?
 # FIXME: EBS volumes?
 
-cluster 'hadoop_demo' do
+ClusterChef.cluster 'demohadoop' do
   cloud :ec2 do
     region              'us-east-1'
     availability_zones  ['us-east-1d']
@@ -63,4 +63,36 @@ cluster 'hadoop_demo' do
   chef_attributes({
       :cluster_size => facet('worker').instances,
     })
+end
+
+ClusterChef.cluster 'democassandra' do
+  cloud :ec2 do
+    region              'us-east-1'
+    availability_zones  ['us-east-1d']
+    flavor              'm1.small'
+    image_name          'mrflip-maverick-client'
+    backing             'ebs'
+    permanent           false
+    elastic_ip          false
+    user_data           :get_name_from => 'broham'
+    spot_price_fraction nil
+    bootstrap_distro    'ubuntu10.04-cluster_chef'
+  end
+
+  role                  "base_role"
+  role                  "chef_client"
+  role                  "ssh"
+  recipe                "cluster_chef::dedicated_server_tuning"
+  role                  "ebs_volumes_attach"
+  role                  "ebs_volumes_mount"
+  role                  "hadoop_s3_keys"
+
+  facet 'datanode' do
+    instances           3
+    cloud.flavor        "m1.large"
+    role                "nfs_client"
+    role                "cassandra_datanode"
+    role                "big_package"
+    role                "hadoop_initial_bootstrap"
+  end
 end
